% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covcor.R
\name{calc_nvars}
\alias{calc_nvars}
\title{Calculate the number of variables implied by a number of pairs}
\usage{
calc_nvars(npairs)
}
\arguments{
\item{npairs}{Count, the number of \strong{pairs}.}
}
\value{
Count, number of \strong{variables}.
}
\description{
Calculate the number of variables implied by a number of pairs.
}
\details{
When creating pairs of variables, e.g. for a correlation/covariance
matrix, the nb of variables \code{nvars} implies a given number of
pairs \code{npairs}. \code{npairs} can be determined since
\code{npairs = binomial{nvars}{2}} which gives
\code{2 x nvars = npairs^2 - npairs} which is a quadratic equation that can
be resolved with \code{npairs = (sqrt(1 - 4 x (-2 x nvars) + 1) / 2}.
}
\examples{
nvars <- calc_nvars(6)
stopifnot(nvars == 4)
}
