% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covcor.R
\name{create_cor}
\alias{create_cor}
\title{Create a correlation matrix, given a vector of correlations}
\usage{
create_cor(cors = 2L, tol = 1e-05)
}
\arguments{
\item{cors}{A vector of correlations of a the nb of variables.}

\item{tol}{Tolerance for the output matrix. Change only if you have a good
reason.}
}
\value{
Correlation matrix with the corelations \code{cors}, or, if
\code{cors} is a count, a random correlation matrix of \code{cors} variables.
}
\description{
Create a correlation matrix, given a vector of correlations. If no
vector of correlations, a random correlation matrix 2 x 2 will be created.
}
\details{
Cretae a positive-definite matrix which represents the correlations given by
the user or, when \code{cors} is a count, a random correlation matrix of
size \code{cors} x \code{cors}.
}
\examples{
Rho <- create_cor(c(0.25, 0.50, 0.75))
stopifnot(identical(dim(Rho), c(3L, 3L)))
Rho <- create_cor(3)
stopifnot(identical(dim(Rho), c(3L, 3L)))
}
