% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covcor.R
\name{create_cor_mat}
\alias{create_cor_mat}
\title{Create a correlation matrix, given a vector of correlations}
\usage{
create_cor_mat(cors = 2L, tol = 1e-06)
}
\arguments{
\item{cors}{A vector of correlations of a the nb of variables.}

\item{tol}{Tolerance for the output matrix. Change only if you have a good
reason.}
}
\value{
Correlation matrix with the correlations \code{cors}, or, if
\code{cors} is a count, a random correlation matrix of \code{cors} variables.
}
\description{
Create a correlation matrix, given a vector of correlations. If no
vector of correlations, a random correlation matrix 2 x 2 will be created.
}
\details{
Create a positive-definite matrix which represents the correlations given by
the user or, when \code{cors} is a count, a random correlation matrix of
size \code{cors} x \code{cors}.
}
\examples{
cors <- c(-0.025667622, -0.009168867, 0.065022678,
 0.189735095, -0.011068128, 0.075086348)
Rho <- create_cor_mat(cors)
stopifnot(identical(dim(Rho), c(4L, 4L)))
Rho <- create_cor_mat(4)
stopifnot(identical(dim(Rho), c(4L, 4L)))
}
\seealso{
calc_nvars sim_cor_mat
}
