% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_intervals.R
\name{gather_intervals}
\alias{gather_intervals}
\title{Gather \code{ggdist::point_interval} summaries (i.e. long format)}
\usage{
gather_intervals(
  data,
  incl = c(".width", ".point", ".interval"),
  excl = "^.+__",
  fun = ggdist::mode_qi,
  ...
)
}
\arguments{
\item{data}{\code{ggdist::point_interval} summaries.}

\item{incl}{Summary variables to include in output.}

\item{excl}{Regex pattern to exclude variables from output.  Usually to
exclude "*__" variables such as "lp__".}

\item{fun}{\code{ggdist::point_interval} function such as \code{mean_qi}.}

\item{...}{Extra arguments used by \code{fun}}
}
\value{
Point and interval summary in long format.
}
\description{
Gather \code{ggdist::point_interval} summaries (i.e. long format).
}
\details{
Take a summary from \code{ggdist::point_interval} and put it in long format
with columns \code{.variable} and .code{.value}.  Grouped variables will
be kept just as the \code{tidybayes::gather_variables()} does it.

This function mimics the behavior of \code{tidybayes::gather_variables()}.
The code is actually pretty much the same.  See @R-tidybayes.
☻
}
\examples{
df <- rbind(
 data.frame(model = "alpha", a = runif(5), b = rnorm(5), c = rexp(5)),
 data.frame(model = "beta", a = runif(5), b = rnorm(5), c = rexp(5))
 )
}
