% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covcor.R
\name{sim_cor_mat}
\alias{sim_cor_mat}
\title{Simulate a correlation matrix from a given nb of variables}
\usage{
sim_cor_mat(nvars = 2, is_cor = TRUE)
}
\arguments{
\item{nvars}{Positive number of variables, must be >= 2.}

\item{is_cor}{TRUE (default): Return the correlation matrix.
FALSE: return the covariance matrix.}
}
\value{
Correlation matrix \code{nvars x nvars}
}
\description{
Simulate a correlation matrix from a given nb of variables.
}
\details{
Simulate a correlation matrix from a given nb of variables by creating
a covariance matrix and transforming it to a correlation matrix with
\code{stats::cov2cor}. The method was copied from \code{simstudy::genCorMat()}
to whom I express my gratitude.
}
\examples{
mat <- sim_cor_mat(4)
stopifnot(identical(dim(mat), c(4L, 4L)))
}
