% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_covcor.R
\name{calc_nvars}
\alias{calc_nvars}
\title{Calculate the number of variables implied by a number of combinations}
\usage{
calc_nvars(ncomb)
}
\arguments{
\item{ncomb}{Count, the number of combinations.}
}
\value{
Count of the number of variables.
}
\description{
Calculate the number of variables implied by a number of combinations..
}
\details{
When creating combinations of variables, e.g. for a correlation/covariance
matrix, the nb of variables \code{nvars} implied a given number of
combinations \code{ncomb} can be deducted since \code{ncomb = binomial{nvars}{2}}
which gives \code{2 x nvars = ncombs^2 - ncomb} which is a quadratic equation
resolved with \code{ncomb = (sqrt(1 - 4 x (-2 x nvars) + 1) / 2}.
}
\examples{
nvars <- calc_nvars(6)
stopifnot(nvars == 4)
}
